ability class and types
